requirements.txt
# ==========================================================
# ✅ Core Testing Dependencies
# ==========================================================
pytest==8.3.3
playwright==1.48.0
pytest-playwright==0.7.1
pytest-xdist==3.8.0
pytest-base-url==2.1.0
pytest-rerunfailures==13.0

# ==========================================================
# ✅ Reporting & Logging
# ==========================================================
allure-pytest==2.13.5
allure-python-commons==2.13.5

# ==========================================================
# ✅ Utilities
# ==========================================================
python-dotenv==1.0.1
requests==2.32.5
typing_extensions>=4.15.0

# ==========================================================
# ✅ Optional (for local HTML reports or logs)
# ==========================================================
# pytest-html==4.1.1
# pytest-metadata==3.1.1
# pytest-sugar==1.1.1
# termcolor==3.1.0
